+++ packages/orq-rc/src/models/operations/listbudgets.ts
@@ -464,7 +464,7 @@ export const ListBudgetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-21T10:57:44.779Z",
+    "2025-10-21T12:55:38.257Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -499,7 +499,7 @@ export const ListBudgetsData$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => ListBudgetsConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-21T10:57:44.779Z"))
+  updated: z.date().default(() => new Date("2025-10-21T12:55:38.257Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {