+++ packages/orq-rc/RELEASES.md
@@ -4419,3 +4419,13 @@ Based on:
 - [typescript v3.14.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.42 - packages/orq-rc
+
+## 2025-10-21 12:57:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.3 (2.727.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.43 - packages/orq-rc